# extensions/jsonrpc/internal/rpcserver/openapi.yaml
openapi: 3.1.0
info:
  title: Sparkdance Managed Lists API
  version: 1.0.0
  description: |
    CRUD API for multi-chain token lists. A list detail is one aggregate document
    containing list-level presentation settings and its managed items. Item routes
    remain available for independent editing. Administrative authentication is
    expected to be enforced by the deployment proxy (for example, Caddy basic_auth).
servers:
  - url: /
...
paths:
  /api/lists:
    get:
      tags: [Lists]
      operationId: listManagedLists
      summary: List all managed-list summaries
      security: [{basicAuth: []}]
      responses:
        '200':
          description: List summaries
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ManagedList'}
        '500': {$ref: '#/components/responses/InternalError'}
    post:
      tags: [Lists]
      operationId: createManagedList
      summary: Create a list
      security: [{basicAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ListCreate'}
      responses:
        '201':
          description: Created list aggregate
          headers:
            Location: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManagedListDocument'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/lists/{listKey}:
    parameters:
      - $ref: '#/components/parameters/ListKey'
    get:
      tags: [Lists]
      operationId: getManagedList
      summary: Get list metadata and all managed items
      security: [{basicAuth: []}]
      responses:
        '200':
          description: Complete management document
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManagedListDocument'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
    put:
      tags: [Lists]
      operationId: replaceManagedList
      summary: Replace a list or create it at this key
      security: [{basicAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ListReplace'}
      responses:
        '200':
          description: Replaced list aggregate
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManagedListDocument'}
        '201':
          description: Created list aggregate
          headers:
            Location: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManagedListDocument'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/InternalError'}
    patch:
      tags: [Lists]
      operationId: patchManagedList
      summary: Modify individual list-level fields
      description: The list key is immutable. Omitted fields are preserved.
      security: [{basicAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ListPatch'}
      responses:
        '200':
          description: Updated list aggregate
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManagedListDocument'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
    delete:
      tags: [Lists]
      operationId: deleteManagedList
      summary: Delete a list and its memberships
      security: [{basicAuth: []}]
      responses:
        '204': {description: Deleted}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/lists/{listKey}/items:
    parameters:
      - $ref: '#/components/parameters/ListKey'
    get:
      tags: [Items]
      operationId: listManagedListItems
      summary: List independently editable items
      security: [{basicAuth: []}]
      responses:
        '200':
          description: Managed items
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ManagedListItem'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
    post:
      tags: [Items]
      operationId: createManagedListItem
      summary: Add a new item to a list
      description: |
        Existing repository metadata is hydrated when chain and address match a
        local asset. A complete ManagedListItem returned by tokenlist, homepage,
        a stablecoin family, or any other list can be posted unchanged here;
        response timestamps are accepted but ignored and regenerated.
      security: [{basicAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ItemCreate'}
      responses:
        '201':
          description: Created item
          headers:
            Location: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ManagedListItem'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Authentication is provided by the deployment proxy.
  parameters:
    ListKey:
      name: listKey
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9._-]*$'
      example: usdt
...
  responses:
    InvalidRequest:
      description: Invalid JSON, unknown field, validation error, or path/body mismatch
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
    ManagedList:
      type: object
      required: [key, enabled]
      properties:
        key: {type: string}
        name: {type: string}
        description: {type: string}
        displayName: {type: string}
        displaySymbol: {type: string}
        logoURI: {type: string, format: uri}
        outputPath: {type: string, description: Portable relative JSON path using forward-slash separators; unique across lists.}
        enabled: {type: boolean}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    ManagedListDocument:
      allOf:
        - $ref: '#/components/schemas/ManagedList'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: {$ref: '#/components/schemas/ManagedListItem'}
    ListCreate:
      type: object
      additionalProperties: false
      required: [key]
      properties:
        key: {type: string, pattern: '^[a-z0-9][a-z0-9._-]*$'}
        name: {type: string}
        description: {type: string}
        displayName: {type: string}
        displaySymbol: {type: string}
        logoURI: {type: string, format: uri}
        outputPath: {type: string, description: Portable relative JSON path using forward-slash separators; unique across lists.}
        enabled: {type: boolean, default: true}
    ListReplace:
      type: object
      additionalProperties: false
      properties:
        key: {type: string, description: If present, must match listKey.}
        name: {type: string}
        description: {type: string}
        displayName: {type: string}
        displaySymbol: {type: string}
        logoURI: {type: string, format: uri}
        outputPath: {type: string, description: Portable relative JSON path using forward-slash separators; unique across lists.}
        enabled: {type: boolean}
    ListPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string}
        description: {type: string}
        displayName: {type: string}
        displaySymbol: {type: string}
        logoURI: {type: string}
        outputPath: {type: string, description: Portable relative JSON path using forward-slash separators; unique across lists.}
        enabled: {type: boolean}
    ManagedToken:
      type: object
      required: [kind, chain, address, decimals, logoExists, hot]
      properties:
        kind: {type: string, enum: [token, native]}
        chain: {type: string}
        chainName: {type: string}
        chainId: {type: integer}
        chainLogoURI: {type: string}
        address: {type: string, description: Empty for native assets.}
        assetId: {type: string}
        type: {type: string}
        name: {type: string}
        symbol: {type: string}
        decimals: {type: integer, minimum: 0, maximum: 255}
        status: {type: string}
        logoURI: {type: string}
        logoExists: {type: boolean}
        explorer: {type: string}
        tags: {type: array, items: {type: string}, uniqueItems: true}
        hot: {type: boolean}
        market: {$ref: '#/components/schemas/ManagedTokenMarket'}
        pairs: {type: array, items: {$ref: '#/components/schemas/TokenPair'}}
        links: {type: array, items: {$ref: '#/components/schemas/Link'}}
    ManagedListItem:
      type: object
      required: [token, enabled, display]
      properties:
        token: {$ref: '#/components/schemas/ManagedToken'}
        slot: {type: string}
        rank: {type: integer, minimum: 0}
        enabled: {type: boolean}
        display: {type: boolean}
        displayName: {type: string}
        displaySymbol: {type: string}
        displayLogoURI: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
...
    ItemCreate:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: {$ref: '#/components/schemas/ManagedTokenInput'}
        slot: {type: string}
        rank: {type: integer, minimum: 0}
        enabled: {type: boolean, default: true}
        display: {type: boolean, default: true}
        displayName: {type: string}
        displaySymbol: {type: string}
        displayLogoURI: {type: string}
        createdAt: {type: string, format: date-time, description: Accepted when copying an item and ignored on write.}
        updatedAt: {type: string, format: date-time, description: Accepted when copying an item and ignored on write.}
...
    ManagedTokenInput:
      type: object
      additionalProperties: false
      required: [chain]
      properties:
        kind: {type: string, enum: [token, native]}
        chain: {type: string}
        chainName: {type: string}
        chainId: {type: integer}
        chainLogoURI: {type: string}
        address: {type: string}
        assetId: {type: string}
        type: {type: string}
        name: {type: string}
        symbol: {type: string}
        decimals: {type: integer, minimum: 0, maximum: 255}
        status: {type: string}
        logoURI: {type: string}
        logoExists: {type: boolean}
        explorer: {type: string}
        tags: {type: array, items: {type: string}}
        hot: {type: boolean}
        market: {$ref: '#/components/schemas/ManagedTokenMarket'}
        pairs: {type: array, items: {$ref: '#/components/schemas/TokenPair'}}
        links: {type: array, items: {$ref: '#/components/schemas/Link'}}
...
    ManagedTokenMarket:
      type: object
      additionalProperties: false
      properties:
        coingeckoId: {type: string}
        marketCapRank: {type: integer, minimum: 0}
        marketCap: {type: number}
        totalVolume: {type: number}
        currentPrice: {type: number}
        lastUpdated: {type: string, format: date-time}
    TokenPair:
      type: object
      additionalProperties: false
      properties:
        base: {type: string}
    Link:
      type: object
      additionalProperties: false
      properties:
        name: {type: string}
        url: {type: string}
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, enum: [invalid_request, not_found, conflict, method_not_allowed, internal_error]}
            message: {type: string}